Define MSE scope, add stream format specs and registries
The scope of the MSE revision is clearly defined to allow the group to consider
additional functionalities strictly limited to that scope.

The Other deliverables section now mentions the maintenance of non-normative
stream format specs and registries for MSE and EME, giving the group the
possibility to define new ones.